Refactor code and change Validator methods to static
