+++ tmtk/toolbox/__init__.py
@@ -1,4 +1,4 @@
 from .generate_chromosomal_regions_file import generate_chromosomal_regions_file
 from .remap_chromosomal_regions import remap_chromosomal_regions
 from . import remap_id
-from .template_interpreter import create_study_from_templates
+from .interpret_templates import create_study_from_templates
